Fix uintmath to work when arrays are uint8-like
Numpy will sometimes throw a typecasting error when casting from higher precision dtypes to uint8. Since the code checks for proper bounds, it shouldn't be a problem.

+++ scikits/image/util/tests/test_uintmath.py
@@ -44,14 +44,14 @@ def test_subtract_array_scalar():
     assert all(subtract_uint(b, a) == d)
 
 def test_multiply_array_scalar():
-    a = np.arange(0, 5, dtype=np.uint8)
+    a = np.arange(0, 5)
     b = 100
     c = (0, 100, 200, 255, 255)
     assert all(multiply_uint(a, b) == c)
     assert all(multiply_uint(b, a) == c)
 
 def test_divide_array_scalar():
-    a = np.arange(0, 10, dtype=np.uint8)
+    a = np.arange(0, 10)
     b = 3
     c = (0, 0, 1, 1, 1, 2, 2, 2, 3, 3)
     assert all(divide_uint(a, b) == c)
